backend/requirements.txt
fastapi
uvicorn
python-multipart
flask
flask-cors
langchain
langchain-community
anthropic
chromadb
python-dotenv
sentence-transformers
pypdf
PyPDF2


